common: break cgo ffmpeg utilities out into their own file
It's good practice to add `import "C"` in files that use C constructs;
this causes CGO_ENABLED=0 to be able to ignore those files and still
compile if other bits aren't necessary.

Needed to use accountmanager as a library in catalyst-api.
